feat(story): Adding camera modal and description from exif
