+++ cmd/fibr/templates/svg.html
@@ -134,3 +134,7 @@
 {{ define "svg-location" }}
   <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 384 512"><path fill="{{ . }}" d="M168.3 499.2C116.1 435 0 279.4 0 192 0 85.96 85.96 0 192 0c106 0 192 85.96 192 192 0 87.4-117 243-168.3 307.2-12.3 15.3-35.1 15.3-47.4 0zM192 256c35.3 0 64-28.7 64-64s-28.7-64-64-64-64 28.7-64 64 28.7 64 64 64z"/></svg>
 {{ end }}
+
+{{ define "svg-camera" }}
+  <svg xmlns="http://www.w3.org/2000/svg" viewBox="0 0 512 512"><path fill="{{ . }}" d="M194.6 32h122.8c20.7 0 39 13.22 45.5 32.82L373.3 96H448c35.3 0 64 28.7 64 64v256c0 35.3-28.7 64-64 64H64c-35.35 0-64-28.7-64-64V160c0-35.3 28.65-64 64-64h74.7l10.4-31.18C155.6 45.22 173.9 32 194.6 32zM256 384c53 0 96-43 96-96 0-53.9-43-96-96-96-53.9 0-96 42.1-96 96 0 53 42.1 96 96 96z"/></svg>
+{{ end }}
